Move Realm URL creation to RealmVerion.
